update benchmark test setup

+++ library/tests/src/test/resources/setupBenchmarkYelp.sh
@@ -1,12 +1,12 @@
-#!/bin/bash
-
-set -euo pipefail
+#!/bin/sh
 
 # NOTE: This requires package mongo-tools (or mongodb-tools) to work.
 
 docker_mongodb_port="3205"
 docker_mongodb_instance="localhost:${docker_mongodb_port}"
-rootdir="$(dirname "$0")/.."
+
+# library/tests/src/test/resources
+rootdir="$(dirname "$0")/../../../../.."
 datadir="$rootdir/data"
 
 source "$rootdir/.env"
@@ -15,10 +15,33 @@ mongosh --port $docker_mongodb_port << EOSQL
     use admin;
     db.auth("${EXAMPLE_USERNAME}", "${EXAMPLE_PASSWORD}");
     use ${BENCHMARK_DATABASE_YELP};
-    db.business.drop()
-    db.user.drop()
-    db.review.drop()
+
+    if (
+        db.business.countDocuments() == 150346 &&
+        db.user.countDocuments() == 1987897 &&
+        db.review.countDocuments() == 6990280
+    ) {
+        exit(42);
+    } else {
+        console.log(db.business.countDocuments());
+        console.log(db.user.countDocuments());
+        console.log(db.review.countDocuments());
+        db.business.drop()
+        db.user.drop()
+        db.review.drop()
+        exit(48);
+    }
 EOSQL
+
+retval=$?
+
+if [ $retval -eq 42 ]; then
+    echo "Yelp (probably) found in MongoDB - skipping import; done."
+    exit 0
+elif [ $retval -ne 48 ]; then
+    echo "Something has gone wrong while checking MongoDB; skipping."
+    exit 1
+fi
 
 mkdir -p "$datadir"
 
@@ -43,9 +66,10 @@ then
     echo "Extracted Yelp data found - skipping unpacking."
 else
     echo "Unpacking..."
-    # export UNZIP_DISABLE_ZIPBOMB_DETECTION=TRUE
+    export UNZIP_DISABLE_ZIPBOMB_DETECTION=TRUE
     unzip "$datadir/Yelp-JSON.zip" -d "$datadir"
     tar --directory "$datadir/Yelp JSON/" -xzf "$datadir/Yelp JSON/yelp_dataset.tar"
+    export UNZIP_DISABLE_ZIPBOMB_DETECTION=
 fi
 
 # WARNING: Importing takes a long time, at least 2 minutes
